+++ src/pages/InteractiveAPI/InteractiveAPI.vue
@@ -38,7 +38,9 @@ query { hello }
         'task_tag_limit',
         'taskTagUsage',
         'task_tag_usage',
-        'mapped_children'
+        'mapped_children',
+        'permissions_info',
+        'users'
       ]
     }
   },